+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ### Unreleased
 
+* (!) Don't immediately start server instances for every already opened file
+* (!) Don't immediately start server instances for newly added workspace folders
 * Dynamically show progress only for the active client workspace
 * Correctly run tasks based on active text editor rather than last opened Rust file
-* (!) Don't immediately try to start server instance for newly added workspace folders
 * Use smooth, universally supported spinner in the status bar ⚙️
 
 ### 0.7.3 - 2020-04-21